Move ApplicationDeployScope to ...internal.app.deploy.pipeline
