+++ cdap-app-fabric/src/test/java/co/cask/cdap/internal/app/deploy/pipeline/ProgramGenerationStageTest.java
@@ -25,7 +25,6 @@ import co.cask.cdap.internal.app.ApplicationSpecificationAdapter;
 import co.cask.cdap.internal.app.Specifications;
 import co.cask.cdap.internal.io.ReflectionSchemaGenerator;
 import co.cask.cdap.internal.pipeline.StageContext;
-import co.cask.cdap.proto.ApplicationDeployScope;
 import co.cask.cdap.test.internal.DefaultId;
 import org.apache.twill.filesystem.LocalLocationFactory;
 import org.apache.twill.filesystem.Location;
